requirements.txt
asgiref==3.5.0
boto3==1.21.21
botocore==1.24.21
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
